requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1

# Herramientas de desarrollo y monitoreo
Flask-DebugToolbar==0.14.1
Werkzeug==3.0.1
gunicorn==21.2.0
psutil==5.9.8
prometheus-client==0.19.0
flask-profiler==1.8

# Documentación
flask-swagger-ui==4.11.1
flask-restx==1.3.0
sphinx==7.2.6
sphinx-rtd-theme==2.0.0 